Exception handling

+++ src/main/java/com/dziombra/dscommerce/services/ProductService.java
@@ -3,15 +3,20 @@ package com.dziombra.dscommerce.services;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.dziombra.dscommerce.dto.ProductDTO;
 import com.dziombra.dscommerce.entities.Product;
 import com.dziombra.dscommerce.repositories.ProductRepository;
+import com.dziombra.dscommerce.services.exceptions.DatabaseException;
 import com.dziombra.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -46,18 +51,29 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update (Long id, ProductDTO dto) {
-		
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);	
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);	
+		}
+		catch(EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado"); 
+		}
 		
 	}
 	
-	@Transactional
-	public void delete (Long id) {
-		repository.deleteById(id);
-
+	@Transactional(propagation = Propagation.SUPPORTS)
+	public void delete(Long id) {
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
